First step

+++ src/lib/functions/postsData.js
@@ -4,15 +4,20 @@ import { editPost }  from '../components/modalPostEdit.js'
 export const savePost = () => {
   const createdPost = document.getElementById('writtePost').value;
   if (createdPost !== '') {
-    return firebase.firestore()
+      return firebase.firestore()
       .collection('posts')
       .add({
         userId: firebase.auth().currentUser.uid,
         userName: firebase.auth().currentUser.displayName,
         text: createdPost,
         privacy: document.getElementById('selectPrivacy').value,
+        postLikes : [],
+        postDislikes :[],
         // Agregar imagenes, etiquetas
         timestamp: firebase.firestore.FieldValue.serverTimestamp()
+      })
+      .then(()=>{
+        window.location.reload();
       })
       .catch(function (error) {
         console.error('Error writing new message to database', error);
@@ -41,15 +46,14 @@ export const showUserPosts = () => {
   .then((querySnapShot) => {
     querySnapShot.forEach((doc) => {
       let arrayUserPosts = doc.data();
-      console.log(arrayUserPosts);
-      console.log(currentUserInfo())
-      const userPostsTemplate = `<div class='post'>
-                                <div class='postUserphoto'><img class='postUserphoto' src="${currentUserInfo()}"></div>
+      const userPostsTemplate = `<div class='postContainer' value="${doc.id}">
+                                <div class='post'>
+                                <div class='postUserphoto'></div>
                                 <div class='postInfo'>
                                 <h2 class='postedUsername'>${arrayUserPosts.userName}
                                 <i id ='editPost' class="fas fa-ellipsis-h">
                                 <div id='menuEdit' class="menuEdit">
-                                    <button id = 'editpost' value='${doc.id}'>Editar</button></a>
+                                    <button id = 'editPostButton' value='${doc.id}'>Editar</button></a>
                                     <button id ='deletepost' value='${doc.id}'>Borrar</button>
                                  </div>
                                  </i>
@@ -59,10 +63,16 @@ export const showUserPosts = () => {
                                   .toDate()
                                   .toDateString()}</p>
                                 <p class='postedText'>${arrayUserPosts.text}</p>
+                                </div>
+                                <div class="postButtons">
+                                <button id="like" class="likeButton">${arrayUserPosts.postLikes.length}<i class="fas fa-heart"></i></button>
+                                <button id="dislike" class="dislikeButton"><i class="fas fa-frown"></i></button>
+                                <button id="comment" class="commentButton"><i class="far fa-comments"></i>Comentar</i></button>
+                                </div>
                                 </div>`;
     userPosts.innerHTML += userPostsTemplate;
   });  
-    const clickEdit = document.querySelectorAll('#editpost');
+    const clickEdit = document.querySelectorAll('#editPostButton');
     clickEdit.forEach(item => {item.addEventListener('click', () => editPost (item.value))});
     const clickDelete = document.querySelectorAll('#deletepost');
     clickDelete.forEach(item => {item.addEventListener('click', () => deletePost (item.value))});
@@ -82,24 +92,14 @@ export const showUserPosts = () => {
 };
 let showMenuEditcontrol = true;
 
-const deletePost= (id) => {
-    firebase.firestore().collection("posts").doc(id).delete().then(() => {
-    console.log("Document successfully deleted! " + id);
-}).catch((error) => {
-  console.error("Error removing document: ", error);
-}
-)};
-
-const currentUserInfo =()=>{
-  let database = firebase
-  .firestore()
-  .collection('userInfo');
-  const holi = database.where( "userId", "==", firebase.auth().currentUser.uid)
-  .get()
-  console.log(holi)
-  /*.then((e) => {
-    e.forEach((doc) => {
-    return doc.data();
+const deletePost = (id) => {
+  console.log('entra a delete' + id);
+    if (confirm("Segura que quieres borrar este post!")) {
+      firebase.firestore().collection("posts").doc(id).delete().then(() => {
+        console.log("Document successfully deleted! " + id);
+        window.location.reload()
+    }).catch((error) => {
+      console.error("Error removing document: ", error);
     })
-  })*/
 }
+};
